+++ client/src/components/mypage/TimeTable/TimeTable.module.css
@@ -1,4 +1,4 @@
 .container {
-  width: 400px;
+  width: 100%;
   position: relative;
 }
