+++ client/src/pages/MyPage/MyPage.jsx
@@ -1,12 +1,10 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import { useRecoilValue } from "recoil";
 import { userState } from "@recoil/user-state";
-import { getProfile, getTimeRecordList, getMyRoom } from "@api/mypage-api";
+import { getProfile, getMyRoom } from "@api/mypage-api";
 import {
   SideNavBar,
-  DatePicker,
-  TimeTable,
-  TimeRecordList,
+  StudyTime,
   MyRoom,
   EditButton,
   MyGoalEditModal,
@@ -15,44 +13,15 @@ import {
 } from "@components/mypage";
 import { Textarea, NicknameEditModal } from "@components/commons";
 import { Layout } from "@components/layout";
-import { getColor, getTimeDiff, getDday, dateParsing, dateFormatting } from "@utils";
+import { getDday, dateParsing } from "@utils";
 import { useAsync, useModal } from "@hooks";
 import styles from "./MyPage.module.css";
 
 function MyPage() {
   const { data: profileData, request: refetchProfile } = useAsync({ requestFn: getProfile, skip: false });
   const { data: myRoomData, request: refetchMyRoom } = useAsync({ requestFn: getMyRoom, skip: false });
-  const [selectedDate, setSelectedDate] = useState(dateFormatting(new Date()));
-  const [timeRecordData, setTimeRecordData] = useState();
-  const [totalStudyTime, setTotalStudyTime] = useState(0);
   const { isModal, openModal, closeModal } = useModal();
   const user = useRecoilValue(userState);
-
-  const fetchTimeRecordData = async (date) => {
-    const response = await getTimeRecordList(date);
-    let total = 0;
-    const array = response.map((data, i) => {
-      const diff = new Date(data.endTime) - new Date(data.startTime);
-      total += diff;
-      return {
-        ...data,
-        startTime: new Date(data.startTime),
-        endTime: new Date(data.endTime),
-        color: getColor(i),
-        studyTime: getTimeDiff(new Date(data.startTime), new Date(data.endTime)),
-      };
-    });
-    setTimeRecordData(array);
-    setTotalStudyTime(total);
-  };
-
-  useEffect(() => {
-    try {
-      fetchTimeRecordData(selectedDate);
-    } catch (error) {
-      console.error(error);
-    }
-  }, [selectedDate]);
 
   const renderModal = (type) => openModal(type);
 
@@ -117,36 +86,7 @@ function MyPage() {
                 </div>
               </div>
             </section>
-            <section className={styles.study_history}>
-              <div className={styles.heading}>공부 기록</div>
-              <div className={styles.contents}>
-                <div className={styles.box}>
-                  <div className={styles.date_box}>
-                    <div className={styles.sub_heading}>날짜</div>
-                    <div className={styles.content}>
-                      <DatePicker onChange={(date) => setSelectedDate(dateFormatting(date))} />
-                    </div>
-                  </div>
-                  <div className={styles.study_time_box}>
-                    <div className={styles.sub_heading}>과목</div>
-                    <div className={styles.content}>
-                      <div className={styles.total_time}>
-                        {`${Math.floor(totalStudyTime / 1000 / 60 / 60)}시간 
-                          ${Math.floor((totalStudyTime / 1000 / 60) % 60)}분 
-                          ${Math.floor(totalStudyTime / 1000) % 60}초`}
-                      </div>
-                      <div className={styles.subject_list}>
-                        <TimeRecordList list={timeRecordData} />
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div>
-                  <div className={styles.sub_heading}>시간표</div>
-                  <TimeTable timeRecordList={timeRecordData} />
-                </div>
-              </div>
-            </section>
+            <StudyTime />
             <section>
               <div className={styles.heading}>
                 <div>생성 스터디룸</div>
