+++ client/src/pages/ShareStudyTime/ShareStudyTime.module.css
@@ -1,0 +1,6 @@
+.share {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  color: white;
+}
